requirements.txt
requests
Pillow
reportlab
